treewide: Staticize dynamic code
The AOT native code compilation function does not support
compiling APIs that use dynamic features, we need to replace
them with static APIs to ensure that the program works properly.

The following APIs have been replaced or modified:
* Json Deserialization
* Assembly Information Read API
* Network Information Extraction API

The network information extraction API does not support dynamic
refreshing of network information after the AOT compilation, so we
can't read the change of port occupancy information after the program
is started, so the current exception determination mechanism is no
longer available, and the program will throw an exception incorrectly
even if the port isn't occupied, so we have changed the exception
determination logic to ensure that the program can work properly.

Signed-off-by: bigsaltyfishes <bigsaltyfishes@gmail.com>